pyproject.toml
[tool.poetry]
name = "protosend"
version = "0.0.0"
description = ""
authors = ["Jesmigel A. Cantos <jesmigel.developer@gmail.com>"]
include = ["src/ska_pst_send"]

[tool.poetry.dependencies]
autoflake = "^1.4"
astropy = "^5.3.1"
black = "^22.1.0"
build = "^0.10.0"
coverage = "^6.3.2"
flake8 = "^6.1.0"
flake8-black = "^0.3.2"
flake8-docstrings = "^1.6.0"
flake8-formatter-junit-xml = "^0.0.6"
flake8-rst-docstrings = "^0.2.5"
isort = "^5.12.0"
mypy = "^1.4.1"
nptyping = "^2.5.0"
pylint = "^2.8.3"
pylint-junit = "^0.3.2"
pytest = "^6.2.0"
pytest-pylint = "^0.18.0"
pytest-cov = "^3.0.0"
python = "^3.10"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
